# api/openapi.yaml
openapi: 3.0.3
info:
  title: Licenses Admin API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /products/{productId}/variants:
    get:
      summary: List variants for product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageProductVariant'
    post:
      summary: Create product variant
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductVariantCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductVariant'
  /variants/{id}:
    get:
      summary: Get variant by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductVariant'
        '404':
          description: Not found
    put:
      summary: Update variant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductVariantUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductVariant'
        '404':
          description: Not found
    delete:
      summary: Delete variant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: No content
...
  /assignments:
    get:
      summary: List assignments
      parameters:
        - in: query
          name: userId
          schema:
            $ref: '#/components/schemas/UUID'
        - in: query
          name: variantId
          schema:
            $ref: '#/components/schemas/UUID'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAssignment'
    post:
      summary: Create assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '409':
          description: Conflict (capacity or duplicate)
  /assignments/{id}:
    get:
      summary: Get assignment by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '404':
          description: Not found
    patch:
      summary: Update assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '404':
          description: Not found
    delete:
      summary: Delete assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: No content
        '404':
          description: Not found
  /manufacturers:
    get:
      summary: List manufacturers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageManufacturer'
    post:
      summary: Create manufacturer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufacturerCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manufacturer'
  /manufacturers/{id}:
    get:
      summary: Get manufacturer by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manufacturer'
        '404':
          description: Not found
    put:
      summary: Update manufacturer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufacturerUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Manufacturer'
        '404':
          description: Not found
    delete:
      summary: Delete manufacturer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: No content
components:
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
  schemas:
    UUID:
      type: string
      format: uuid
...
    ProductVariant:
      type: object
      required: [id, productId, key, name]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        productId: { $ref: '#/components/schemas/UUID' }
        key: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        price:
          type: number
          format: double
          nullable: true
    ProductVariantCreateRequest:
      type: object
      required: [key, name]
      properties:
        key: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        price:
          type: number
          format: double
          nullable: true
    ProductVariantUpdateRequest:
      type: object
      properties:
        key: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        price:
          type: number
          format: double
          nullable: true
    PageProductVariant:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/ProductVariant' }
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
...
    Assignment:
      type: object
      required: [id, userId, productVariantId]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        userId: { $ref: '#/components/schemas/UUID' }
        productVariantId: { $ref: '#/components/schemas/UUID' }
        note:
          type: string
          nullable: true
    AssignmentCreateRequest:
      type: object
      required: [userId, productVariantId]
      properties:
        userId: { $ref: '#/components/schemas/UUID' }
        productVariantId: { $ref: '#/components/schemas/UUID' }
        note: { type: string, nullable: true }
    AssignmentUpdateRequest:
      type: object
      properties:
        _unused:
          type: string
          nullable: true
    PageAssignment:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/Assignment' }
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
    Manufacturer:
      type: object
      required: [id, name]
      properties:
        id: { $ref: '#/components/schemas/UUID' }
        name: { type: string }
        url: { type: string, nullable: true }
        description: { type: string, nullable: true }
    ManufacturerCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        url: { type: string, nullable: true }
        description: { type: string, nullable: true }
    ManufacturerUpdateRequest:
      type: object
      properties:
        name: { type: string }
        url: { type: string, nullable: true }
        description: { type: string, nullable: true }
    PageManufacturer:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/Manufacturer' }
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }