# target/generated-sources/openapi/src/main/java/io/meimberg/licenses/web/api/AssignmentsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.meimberg.licenses.web.api;

import io.meimberg.licenses.web.dto.Assignment;
import io.meimberg.licenses.web.dto.AssignmentCreateRequest;
import io.meimberg.licenses.web.dto.AssignmentUpdateRequest;
import io.meimberg.licenses.web.dto.PageAssignment;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-23T15:50:51.445317955+01:00[Europe/Berlin]", comments = "Generator version: 7.7.0")
@Validated
@Tag(name = "assignments", description = "the assignments API")
public interface AssignmentsApi {

    /**
     * GET /assignments : List assignments
     *
     * @param userId  (optional)
     * @param variantId  (optional)
     * @param page  (optional)
     * @param size  (optional)
     * @param sort  (optional)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "assignmentsGet",
        summary = "List assignments",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PageAssignment.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/assignments",
        produces = { "application/json" }
    )
    
    ResponseEntity<PageAssignment> assignmentsGet(
        @Parameter(name = "userId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "userId", required = false) UUID userId,
        @Parameter(name = "variantId", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "variantId", required = false) UUID variantId,
        @Min(0) @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false) Integer page,
        @Min(1) @Max(200) @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false) Integer size,
        @Parameter(name = "sort", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) String sort
    );


    /**
     * DELETE /assignments/{id} : Delete assignment
     *
     * @param id  (required)
     * @return No content (status code 204)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "assignmentsIdDelete",
        summary = "Delete assignment",
        responses = {
            @ApiResponse(responseCode = "204", description = "No content"),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/assignments/{id}"
    )
    
    ResponseEntity<Void> assignmentsIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * GET /assignments/{id} : Get assignment by id
     *
     * @param id  (required)
     * @return OK (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "assignmentsIdGet",
        summary = "Get assignment by id",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Assignment.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/assignments/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Assignment> assignmentsIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    );


    /**
     * PATCH /assignments/{id} : Update assignment
     *
     * @param id  (required)
     * @param assignmentUpdateRequest  (required)
     * @return OK (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "assignmentsIdPatch",
        summary = "Update assignment",
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Assignment.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/assignments/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Assignment> assignmentsIdPatch(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "AssignmentUpdateRequest", description = "", required = true) @Valid @RequestBody AssignmentUpdateRequest assignmentUpdateRequest
    );


    /**
     * POST /assignments : Create assignment
     *
     * @param assignmentCreateRequest  (required)
     * @return Created (status code 201)
     *         or Conflict (capacity or duplicate) (status code 409)
     */
    @Operation(
        operationId = "assignmentsPost",
        summary = "Create assignment",
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Assignment.class))
            }),
            @ApiResponse(responseCode = "409", description = "Conflict (capacity or duplicate)")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/assignments",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Assignment> assignmentsPost(
        @Parameter(name = "AssignmentCreateRequest", description = "", required = true) @Valid @RequestBody AssignmentCreateRequest assignmentCreateRequest
    );

}